Update file moteur.py

+++ code_raspberry/moteur.py
@@ -20,7 +20,7 @@ def avance_corrige(moteur_princ, ratio, vitesse):
     """ Fait avancer le robot en imposant un ratio (entre 0 et 1) entre les vitesses des moteurs, 
         le moteur le plus rapide est moteur_prin entre left et right """
 
-    ratio = int(ratio)
+    ratio = float(ratio)
     vitesse = int(vitesse)
 
     if moteur_princ == "left":
